IMPALA-8669 Support giving a terminal commit in compare_branches.py
Currently, bin/compare_branches.py (used in cherrypick-2.x-and-test
Jenkins job) will cherry-pick as more commits as possible. However, a
clean pick from the master branch does not mean it can always pass the
tests. If we find such a problematic commit, we'd like to let the
cherrypick-2.x-and-test Jenkins job cherry-pick till the commit before
it. And then submit only a patch for the problematic commit for review.

This patch adds a new parameter "source_terminal_commit" for specifying
where to stop. It's the full commit hash of the terminal commit in the
source branch. bin/compare_branches.py will stop after cherry-picking
this commit.

Tests:
 - Test locally by pushing the 2.x branch to my repo ("self") and run
   bin/compare_branches.py --cherry_pick --partial_ok \
     --source_remote_name asf-gerrit --target_remote_name self -v \
     --source_terminal_commit def5c881b1dee93848daea52d93a0602e11338ca

Change-Id: I7a4dfae26f8f616c6f550114edfbafd6fc1c068d
Reviewed-on: http://gerrit.cloudera.org:8080/13660
Reviewed-by: Tim Armstrong <tarmstrong@cloudera.com>
Tested-by: Quanlong Huang <huangquanlong@gmail.com>