IMPALA-8443: Record time spent in authorization in the runtime profile
The analysis and authorization is handled together as authorization
depends on the results of the analysis. The timeline EventSequence is
moved into the AuthorizationContext and the markEvent is called during
the postAuthorize method.

In some cases the EventSequence is not available when the
AuthorizationContext is created therefore it is wrapped in Optional.

Change-Id: I5bb85e57fcc75d41f3eb2911e6d375e0da6f82ae
Reviewed-on: http://gerrit.cloudera.org:8080/13353
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>