IMPALA-7734: Catalog and Statestore memz page shows useless memory
The MemTracker is removed from StatestoreD main and CatalogD main as
it is not used. The /memz page related metrics are refactored to a
a separate file. When the MemTracker is not available during callback
initialization it should not be part of the callback either. Therefore,
when MemTracker is not part of the callback it is not displayed.

Before/after pictures of the /memz page can be found in the Jira.

Change-Id: I1a42434b98aec1d9e0be6ae52325049378fdde23
Reviewed-on: http://gerrit.cloudera.org:8080/13670
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>