IMPALA-7486: Add specialized estimation scheme for dedicated coordinators
This patch computes two memory estimates in the frontend:
an estimate for any host that is an executor (including
a combined coordinator and executor) and an estimate
for a dedicated coordinator. This is computed regardless
of whether it is a dedicated coordinator or not.

Admission control then, in the case when the coordinator
is dedicated, uses the coordinator memory estimate for
the coordinator node and the executor memory estimate
for all other nodes.

Other highlights:
 - if MEM_LIMIT query option is set, it is applied to all backends,
   both executors and coordinators.
 - the min_query_mem_limit pool config is not enforced on the
   dedicated coordinator estimates unless MEM_LIMIT query option is set.
 - the lower cap on estimates and the admission checks based on the
   min mem limit required for reservation are applied separately on
   coordinator's and executors' mem requirements.
 - Added a hidden startup option 'use_dedicated_coordinator_estimates'
   which if set to false, reverts to previous estimation behavior.

Testing:
- Added unit test for admission/rejection in dedicated
  coordinator clusters.
- Added end to end tests.

Change-Id: I2b94e7293b91dec8a18491079c34923eadd94b21
Reviewed-on: http://gerrit.cloudera.org:8080/13740
Reviewed-by: Bikramjeet Vig <bikramjeet.vig@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>