IMPALA-8565: larger query in test_admission_controller
It appears that in some rare cases, fragments from the
query used can finish too early, before the client has
started fetching to hit EOS.

Reduce the chances of this happening by increasing
the query size and reducing the fetch frequency.

Testing:
I could not reproduce this rare flake locally, probably
because my hardware is different from the hardware that
the failures occurred on.

I looped the failing test for a while locally to try and
ensure that the flakiness did not at least get worse.

Change-Id: Ifd7716d5d5ccc4b2f79b3acf2fbc880716b4c0fe
Reviewed-on: http://gerrit.cloudera.org:8080/13718
Reviewed-by: Bikramjeet Vig <bikramjeet.vig@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>