IMPALA-5031: method calls on NULL are not UBSAN-clean
According to [expr.post] in the C++14 standard, a call to a member
function like a->b() is interpreted as (a->b)(). In other words, the
dereferencing is done separately from the call. This makes calling
member functions on nullptr undefined behavior, since the dereference
invokes undefined behavior.

This fixes such an error in exec-node.cc in the end-to-end tests. The
interesting part of the backtrace is:

exec/exec-node.cc:396:27: runtime error: member call on null pointer
  of type 'MemTracker'
    #0 in ExecNode::ExecDebugActionImpl(TExecNodePhase::type,
       RuntimeState*) exec/exec-node.cc:396:27
    #1 in ExecNode::ExecDebugAction(TExecNodePhase::type,
       RuntimeState*) exec/exec-node.h:379:12
    #2 in ExecNode::Prepare(RuntimeState*) exec/exec-node.cc:106:43
    #3 in TopNNode::Prepare(RuntimeState*) exec/topn-node.cc:75:53

Change-Id: Id62d1c504a273451dc1be6831a473f6c7115b403
Reviewed-on: http://gerrit.cloudera.org:8080/13769
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>