Bump CDP_BUILD_NUMBER to 1235229
The newer CDP build includes fix for HIVE-21932 which is required to fix
test failures on CDP builds.

Also, this patch fixes two test failures on CDP.

1 Fixes the MetastoreEventsProcessorTest on CDP caused due to the
serialization difference of the event messages between CDH and CDP
builds.
2. Fixes the Ranger audit test failure caused due to the missing cluster
name paramter from the Audit events from Ranger. This is needed a change
in the ranger-hive-audit.xml due to RANGER-2458

Testing:
1. Build against CDP and CDH builds
2. Run the previously failing tests on CDP which were caused by
HIVE-21932
3. Run full tests with CDP and CDH builds

Change-Id: I6d0fa856fa6c7cb1f4c9ee0e4250cfa1e8a14595
Reviewed-on: http://gerrit.cloudera.org:8080/13799
Reviewed-by: Zoltan Borok-Nagy <boroknagyz@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>