IMPALA-5149: Provide query profile in JSON format
Description:
Today there is a text and Thrift version of the query profile, but
it would be useful to have a JSON version for portability and machine
consumption. The ultimate goal is to have a Download JSON format
profile link along with the other two formats.

Modification:
1.Add Json format download option in impalad profile page
2.Add ToJson() function for RuntimeProfile, Counters, EventSequence
3.Add JSON format into QueryStateRecord
4.Add tests for E2E test, unit test, hs2 test
5.Modify query profile page to different download option
6.Modify HS2 server to support get JSON format profile

Future compatibility:
The schema of the JSON format can be changed in the future with
the standardization of Profile and Counter structure.

Tests:
E2E tests:
tests/webserver/test_web_pages.py - test_download_profile
 * merge text and json format download together
HS2 tests:
tests/hs2/test_hs2.py - test_get_profile
 * add json format test
BE Unit tests:
be/src/util/runtime-profile-test.cc
 * ToJson.RuntimeProfileToJsonTest
 * ToJson.EmptyTest
 * ToJson.EventSequenceToJsonTest
 * ToJson.TimeSeriesCounterToJsonTest

Change-Id: I8181ac818bf22207ca1deabd9220c397ae723ec1
Reviewed-on: http://gerrit.cloudera.org:8080/13801
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>