IMPALA-6663: Expose current DDL metrics on WebUI
This change adds a new '/operations' tab to the Catalog WebUI. On  this
page the users can see two summary tables, the first report is based on
how many catalog operations are running and which tables are involved.
The second report shows the number of catalog operations on table level.

A new 'monitor' package is created to collect the monitoring classes.
Each DDL/FinalizeDML/ResetMetada request get a class with common base
class. The output of these classes are summarized and can be accessed
with a new JNI call.

A screenshot of the new '/operations' page can be found in the Jira.

Testing:
Unit tests were added to test the counters and a web server test to
verify the availability of the page.

Change-Id: I0ed76f134bad6d3b3d4dce132365a53a01e9512a
Reviewed-on: http://gerrit.cloudera.org:8080/13806
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>