IMPALA-8751: Skip failing Kudu - HMS integration tests with Hive 3
These tests broke with the newer versions of CDP Hive, probably
because it started to send gzipped notifications, which lead to
errors like this in Kudu:
Not implemented: unknown message format: gzip(json-2.0)

The Kudu side fix does not seem trivial, so I am skipping these
tests for now. Generally Kudu is not tested well with Hive 3 yet,
so we cannot assume HMS notifications to work.

Change-Id: Ic19b8d93eaa6e8ec886c5704578563fb0871f941
Reviewed-on: http://gerrit.cloudera.org:8080/13854
Reviewed-by: Gabor Kaszab <gaborkaszab@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>