IMPALA-8766: Change cloud dependencies to use hadoop-cloud-storage
Currently, fe/pom.xml includes direct dependencies on hadoop-aws,
hadoop-azure, and hadoop-azure-datalake. Hadoop has a meta-package
hadoop-cloud-storage that provides these dependencies with appropriate
exclusions. This would not change Impala's build significantly, but
it allows different versions or distributions of Hadoop to add
necessary runtime jars to the classpath. For example, distributions
of Hadoop that use Apache Knox for identity management can add
Knox jars to this meta-package.

This switches Impala to use the hadoop-cloud-storage package. This
does not change the default configuration, but the USE_CDP_HIVE=true
configuration gains Apache Knox jars.

Due to the structure of the CDP maven repository, it has dependencies
on versions of jars that are not in the CDP maven repository. So,
USE_CDP_HIVE=true requires the Hortonworks public maven repository.
This is added, but it is only enabled with USE_CDP_HIVE=true and only
for releases (no SNAPSHOTs).

Testing:
 - Ran core tests
 - Manually verified classpath changes

Change-Id: I08f1c36ecf54ac277d99e2d2843163eada732e50
Reviewed-on: http://gerrit.cloudera.org:8080/13872
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Reviewed-by: Vihang Karajgaonkar <vihang@cloudera.com>