IMPALA-8516: Fix the sha512sum check for the Maven download
This is a follow-up fix for commit 327b938214:
When bin/bootstrap_system.sh downloads Maven, it performs a checksum
check on the downloaded archive before expanding it. The check is
performed using sha512sum, which has quite particular requirements
for its inputs:

    When checking, the input should be a former output of this program.
    The default mode is to print a line with checksum, a character
    indicating input mode ('*' for binary, space for text),
    and name for each FILE.

In our case the checksum lines ended up having two space characters
between the checksum and the file name. Unfortunately, when the
original commit copied the maven download-and-install stanza to
the new location, the second whitespace was dropped, which made this
line fail on CentOS 6, where the sha512sum implementation takes the
above quoted requirement literally.

This patch restores the extra whitespace, fixing Maven installation
for CentOS 6.

Change-Id: Ic8aa4d78b34d60a01fc8f6afc336d66ee5a005cb
Reviewed-on: http://gerrit.cloudera.org:8080/13876
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>