Add default value to switch in kudu-util
KuduDataTypeToColumnType in kudu-util.cc has a switch that translates
kudu::client::KuduColumnSchema::DataType to PrimitiveType. It returns
PrimitiveType::INVALID_TYPE if an invalid type is passed, but it was
done after the switch instead of the 'default' case which leads to a
warning if a new unhandled Kudu type is added to the header.

Change-Id: I81a8ab4b912d443582c8ac3899f370a4b838e934
Reviewed-on: http://gerrit.cloudera.org:8080/13901
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>