IMPALA-8810: Fix ExplainTest when Orc support is disabled
The problem was that ExplainTest only mocks
HdfsPartition.getFileFormat() but not
HdfsPartition.getInputFormatDescriptor(), and the latter is called
then dereferenced during testScanNodeFsScheme() (only if Orc support
is disabled), leading to null pointer exception.

I went for the lazy solution of replacing
getInputFormatDescriptor().getFileFormat() calls with getFileFormat().
The other solution would have been to mock getInputFormatDescriptor()
too, but I think that would have only added unnecessary dependencies to
the test.

Change-Id: I0d8bcb67e095a63beb9059568ba34b7fd261f233
Reviewed-on: http://gerrit.cloudera.org:8080/13956
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>