IMPALA-8637: Implement transaction handling and locking for ACID queries
Adds background thread to the Frontend that periodically heartbeats
the opened transactions and locks. Most of the logic is implemented
in the newly added TransactionKeepalive class.

TransactionKeepalive keeps track of the creation time of the
transactions and locks, and only heartbeat them if they are older
then the sleep interval. This way we don't heartbeat short-running
queries unnecessarily.

Testing:
Added an exhaustive test that checks heartbeating. It creates a long
running transaction with sleep(). While the transaction is executing
the test periodically checks whether there is an OPEN transaction that
has sent a heartbeat to HMS. If it found one then the test succeeds.

TODOs:
 * add metrics

Change-Id: Iaa37899b24aa114be642bf8772b4e0f882865cfa
Reviewed-on: http://gerrit.cloudera.org:8080/13968
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>